+++ KPCF/changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 
+
+
 ## [1.1.0] - 2018-11-05
 ### Added
 - Alphabetical sorting for item lists
