+++ KPCF/config.sqf
@@ -1,11 +1,14 @@
 /*
-    Killah Potatoes Cratefiller v1.1.0
+    KP Cratefiller config file
 
+    File: config.sqf
     Author: Dubjunk - https://github.com/KillahPotatoes
+    Date: 2019-05-09
+    Last Update: 2019-05-09
     License: GNU General Public License v3.0 - https://www.gnu.org/licenses/gpl-3.0.html
 
     Description:
-    Configuration file for various variables of the KP cratefiller.
+        This file contains several configs for the cratefiller.
 */
 
 // The Base object is for the player interaction, so at these object you can open the dialog
