+++ KPCF/fnc/fn_addEquipment.sqf
