+++ KPCF/fnc/fn_createEquipmentList.sqf
