+++ KPCF/fnc/fn_createSubList.sqf
