+++ KPCF/fnc/fn_deleteCrate.sqf
