+++ KPCF/fnc/fn_deletePreset.sqf
