+++ KPCF/fnc/fn_getConfigPath.sqf
