+++ KPCF/fnc/fn_getInventory.sqf
