+++ KPCF/fnc/fn_getItems.sqf
