+++ KPCF/fnc/fn_getNearStorages.sqf
