+++ KPCF/fnc/fn_init.sqf
