+++ KPCF/fnc/fn_manageAceActions.sqf
