+++ KPCF/fnc/fn_manageActions.sqf
