+++ KPCF/fnc/fn_openDialog.sqf
