+++ KPCF/fnc/fn_removeEquipment.sqf
