+++ KPCF/fnc/fn_setActiveStorage.sqf
