+++ KPCF/fnc/fn_setInventory.sqf
