+++ KPCF/fnc/fn_showInventory.sqf
