+++ KPCF/fnc/fn_showPresets.sqf
