+++ KPCF/fnc/fn_sortList.sqf
