+++ KPCF/fnc/fn_spawnCrate.sqf
