+++ KPCF/variables.sqf
