+++ KPGUI/res/icon_recyclebin.paa
