+++ KPGUI/res/icon_refresh.paa
