+++ KPGUI/res/kp512_CA.paa
