+++ KP_functions.hpp
@@ -1,0 +1,16 @@
+/*
+    KP function library
+
+    File: KP_functions.hpp
+    Author: Dubjunk - https://github.com/KillahPotatoes
+    Date: 2019-05-09
+    Last Update: 2019-05-09
+    License: GNU General Public License v3.0 - https://www.gnu.org/licenses/gpl-3.0.html
+
+    Description:
+        Fetches all functions from the modules.
+*/
+
+class KP {
+    #include "KPCF\functions.hpp"
+};
