+++ KP_ui.hpp
@@ -1,0 +1,15 @@
+/*
+    KP ui defines
+
+    File: KP_ui.hpp
+    Author: KP Liberation Dev Team - https://github.com/KillahPotatoes
+    Date: 2017-10-16
+    Last Update: 2019-05-09
+    License: GNU General Public License v3.0 - https://www.gnu.org/licenses/gpl-3.0.html
+
+    Description:
+        Includes all ui defines, dialogs and elements from each module.
+*/
+
+#include "KPGUI\KPGUI_defines.hpp"
+#include "KPCF\ui\KPCF_dialog.hpp"
