+++ description.ext
@@ -9,10 +9,12 @@
     Otherwise copy the content below this comment block right at the end of your description.ext file, if you're unsure how to merge them.
 */
 
-#include "KPGUI\KPGUI_defines.hpp"
-#include "KPCF\ui\KPCF_dialog.hpp"
+/*
+    ----- Config Includes -----
+*/
 
 class CfgFunctions {
-    // If you already have this class opened in your description.ext, copy and paste just this include inside of it.
-    #include "KPCF\KPCF_functions.hpp"
+    #include "KP_functions.hpp"
 };
+
+#include "KP_ui.hpp"
